+++ Changes.md
@@ -3,6 +3,8 @@
 - refactor ./config.js as an es6 class
 - update README syntax and improve formatting
 - use path.resolve instead of ./dir/file (2x)
+- watch: recursive=true
+- permit retrieval of fully qualified path
 
 ## 1.0.16 - 2018-11-02